requirements.txt
Django==4.0.4
psycopg2==2.9.3
black==22.3.0
djangorestframework==3.13.1
requests[security]==2.27.1
unittest2==1.1.0
